+++ changelog.md
@@ -10,6 +10,7 @@
 * More descriptive status messages
 * Read/write DICOM overlay
 * Cancel label selection
+* Split tissue split tissue near to missed branche between selected branches
 
 ## 1.9.51 2018-07-21